requirements.txt
future
emoji
beautifulsoup4
telegraph
requests
sqlalchemy
tswift
#coffeehouse
tldextract
gtts
python-telegram-bot==11.1.0
numpy
urbandict
pyowm
wikipedia
kbbi
feedparser
alphabet_detector
psycopg2
feedparser
#telegram
pynewtonmath
#memes vvv
spongemock
zalgo-text
geopy
nltk
psutil
bleach
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
#translator
googletrans
jikanpy
speedtest-cli
pythonping
